# swagger.yaml
swagger: "2.0"
info:
  description: "API para gerenciamento de usuários."
  version: "1.0.0"
  title: "API Usuários"
basePath: "/"
tags:
  - name: "Usuários"
    description: "Operações relacionadas a usuários"
schemes:
  - "http"
paths:
  /cadastrar_usuario:
    post:
      tags:
        - "Usuários"
      summary: "Cadastrar Usuário"
      description: "Cadastra um novo usuário com nome e email."
      consumes:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Nome e email do usuário"
          required: true
          schema:
            type: "object"
            required:
              - "nome"
              - "email"
            properties:
              nome:
                type: "string"
              email:
                type: "string"
      responses:
        201:
          description: "Usuário cadastrado com sucesso"
        400:
          description: "Erro na requisição"
  /buscar_usuario:
    get:
      tags:
        - "Usuários"
      summary: "Buscar Usuários"
      description: "Retorna todos os usuários cadastrados."
      responses:
        200:
          description: "Lista de usuários cadastrados"
          schema:
            type: "array"
            items:
              type: "object"
              properties:
                id:
                  type: "integer"
                nome:
                  type: "string"
                email:
                  type: "string"
  /deletar_usuario/{id}:
    delete:
      tags:
        - "Usuários"
      summary: "Deletar Usuário"
      description: "Deleta um usuário pelo ID."
      parameters:
        - in: "path"
          name: "id"
          description: "ID do usuário"
          required: true
          type: "integer"
      responses:
        200:
          description: "Usuário deletado com sucesso"
        404:
          description: "Usuário não encontrado"